Fix currently broken steamykitchend and closetcooking tests, as bs4 was updated.
Add a backward compatible code for older versions of bs4 where the correct meta
tag is also found.

Upgrade get_minutes function to be able to work if simple text (NavigableString)
is given to it. (So now it works both with Tags/Html Nodes and simple strings
as a result