Initial work to return grouped ingredients from a recipe.
Add ingredient_groups function to AbstractScraper. By default, this returns a list containing a single IngredientGroup object, with purpose set to None and ingredients set to self.ingredients(). This function can be overridden in site specific classes derived from AbstractScraper to return the ingredient grouped according to the recipe.

Add _grouping_utils.py which contains some helper functions for grouping ingredients. This is where the IngredientGroup dataclass is defined. This also includes a helper function that will group the ingredients for a recipe if the CSS selectors for the group heading and group elements are provided.