Add test_consistent_ingredient_lists test to ScraperTest base class. This test checks that the ingredients in the list of grouped ingredients are the same as the ingredients returned from .ingredients(). This test is inherited by all scraper test class derived from ScraperTest.
The addition of this test function to the ScraperTest class means the default unittest test loader will try to execute the test on the ScraperTest class, which will fail. To resolve this, we define a load_tests function that overrides the default unittest test loader and removes any discovered tests that have a type of ScraperTest, leaving just the tests we actually want to run.