Add test_consistent_ingredient_lists test to ScraperTest base class. This test checks that the ingredients in the list of grouped ingredients are the same as the ingredients returned from .ingredients(). This test is inherited by all scraper test class derived from ScraperTest.
This requires additional updates to the ScraperTest base class to prevent test_* functions running on ScraperTest class directly. The default unittest test runner will run all test_* functions on all discovered classes including base classes. We don't want to run any tests on ScraperTest, only on the derived classes, so we override the run function to bypass any ScraperTest instances. This also updates the setUpClass class method so that it doesn't raise an error if called on ScraperTest instances.