Ingredient groups (#781)
* Initial work to return grouped ingredients from a recipe.

Add ingredient_groups function to AbstractScraper. By default, this returns a list containing a single IngredientGroup object, with purpose set to None and ingredients set to self.ingredients(). This function can be overridden in site specific classes derived from AbstractScraper to return the ingredient grouped according to the recipe.

Add _grouping_utils.py which contains some helper functions for grouping ingredients. This is where the IngredientGroup dataclass is defined. This also includes a helper function that will group the ingredients for a recipe if the CSS selectors for the group heading and group elements are provided.

* Update NIHHealthyEating scraper to use definition of IngredientGroup from _grouping_utils.py.
Update PickUpLimes scraper to use group_ingredients helper function to return grouped ingredients.

* Update PickUpLimes test to test ingredient groups.
Update test_utils to fix test now there is a new class function.

* Correctly serialise IngredientGroup dataclass into a dict when classing the .to_json() method on a scraper object.

* Modify bbcgoodfood, budgetbytes and rainbowplantlife scrapers to support reutrning grouped ingredients.
Update existing tests and add new tests to cover recipes with and without grouped ingredients.

* Add tests for _grouping_utils helper functions.

* Add ingredient_groups() function to list in README

* Add type hint to ingredient_group function of AbstractScraper

* Replace test recipe for PickUpLimes scraper.
The previous test recipe was missing ingredients from the ld+json which was making the new .ingredient_groups() inconsistent with .ingredients()

* Add test_consistent_ingredient_lists test to ScraperTest base class. This test checks that the ingredients in the list of grouped ingredients are the same as the ingredients returned from .ingredients(). This test is inherited by all scraper test class derived from ScraperTest.

The addition of this test function to the ScraperTest class means the default unittest test loader will try to execute the test on the ScraperTest class, which will fail. To resolve this, we define a load_tests function that overrides the default unittest test loader and removes any discovered tests that have a type of ScraperTest, leaving just the tests we actually want to run.

* Revert "Add test_consistent_ingredient_lists test to ScraperTest base class. This test checks that the ingredients in the list of grouped ingredients are the same as the ingredients returned from .ingredients(). This test is inherited by all scraper test class derived from ScraperTest."

This reverts commit 8d334f1ac102837c3060cb48d48258b8757e2ebc.

* Move ScraperTest class definition inside another class (AbstractTest).
This is done to hide the test_* functions of ScraperTest from the unittest test runner, so the test_* functions are only called by classes derived from AbstractTest.ScraperTest. Without this, unittest will attempt to run ScraperTest.test_*, which will fail due to the setUpClass failing when not called on a derived class.

* Revert "Move ScraperTest class definition inside another class (AbstractTest)."

This reverts commit 40cf9a91a434d1f02817b062e24d7e5066f562f7.

* Add test_consistent_ingredient_lists test to ScraperTest base class. This test checks that the ingredients in the list of grouped ingredients are the same as the ingredients returned from .ingredients(). This test is inherited by all scraper test class derived from ScraperTest.

This requires additional updates to the ScraperTest base class to prevent test_* functions running on ScraperTest class directly. The default unittest test runner will run all test_* functions on all discovered classes including base classes. We don't want to run any tests on ScraperTest, only on the derived classes, so we override the run function to bypass any ScraperTest instances. This also updates the setUpClass class method so that it doesn't raise an error if called on ScraperTest instances.

* Add missing description to docstring for output of score_sentence_similarity.