ci: fixup / cleanup: remove 'setuptools' search line from publish workflows.
Explanation: our CI now runs the publish workflows on Python 3.12, and in that version of Python, setuptools is not installed to virtualenvs -- such as the one used by the 'build' command that constructs our library's package -- by default.  Since 'setuptools' isn't present, a grep for it in the output of 'pip list' will fail with a non-zero exit code, terminating the workflow job with an error status.  We don't want that, so remove the relevant search (grep) command from the version-output step.