cleanup: remove character-to-token expansion ('h' to 'hours') that already seems to be covered by a corresponding regex match for a single 'h' character
This code branch _does_ get followed sometimes during unit testing, so it's not dead code.  I think it's intended to ensure that a singleton 'h' char is matched by TIME_REGEX.. but that regex already contains a pattern that matches 'h' or 'hours' equivalently, so I believe that the expansion is redundant.