Add the 'description' field for seven existing parsers
See [a downstream PR for more context](https://github.com/poundifdef/plainoldrecipe/pull/176), which is where I came across this.

This adds the description field for a few parsers that definitely support it. These are only ones in-common with the downstream app linked above, there are likely plenty more that could benefit from this.

This is broadly similar to #554